requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
bcrypt==4.1.2
pyjwt==2.8.0
pydantic[email]==2.5.0
python-dotenv==1.0.0

